Glow functionality
Glow values are validated at 1 or 0 (glow or no glow).  Added icon in shard creation/edit form to select glow state.  Glow state is passed from the shard to the form, changed in the form along with text and tint, and then passed from the form to the create and edit helper functions.  Glow state is also passed from the form to the front-end shard object.  In the .ejs shardList partial, a class of "glow" is added if the data-shard-glow equals 1.  CSS selects for glow class pseudo-element and adds box shadow and background.