Spark functionality
Added column called "spark" to shards table.  A spark is the inspiration for a shard.  Added randomSpark function to shards.js to select random shard from array.  Random spark text is displayed in the create shard form, along with a button that refreshes a new random spark.  On form submission, the spark is pushed to the spark controller and model to add the spark text to that shard's new row.  Spark span added to shard list that is rendered from shard object properties.     Also the main animation speed property was made faster.