Re-factored for client-side rendering of shards
Because the on-screen shard elements are now generated dynamically on the front end with a library, I refactored the shards.js code for client-side rendering.  The backend now sends a JSON file with an array of shard objects.  This array is used to generate the colorful "kaleidoscope" of shards on the front end.  I also created closures to store states for the shards, voronoi, and the app in general.  The UI is controlled more broadly by switching the app's state from "mainView" to "shardEditing," for example.  Rather than updating the UI piecemeal in each event listener, the app state is changed, which then updates the UI accordingly.