Prompt generation
Since an SLM running in the browser wasn't powerful enough to create a personality analysis from the user's questions, the model was moved to the backend.  GPT 5 nano is now used on the backend instead of GPT 2 running in the browser.  Also revised the sculptures script file for more clarity on which functions are CRUD requests vs helper or event handler functions